Basic framework for supporting geoasn whitelists
We want the prober to produce a "0" result for any geo-asn
pairs that are in the whitelist but are not in the probelist.

This doesn't work yet because I haven't actually written the
code to read the whitelist and create time series keys for each
entity yet...

This approach ensures that we will always produce a valid TS
value for each whitelisted pair, even if there are no responsive
/24s (as opposed to an empty data point which could be confused
with a measurement failure).